+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v2.0.0 (Unreleased)
 
 ### Added
+  - Admin / Moderator actions are now logged and available for viewing in Command & Control
   - {site_name} is now a default template var
   - Allow { and } to be escaped in templates by using {{ and }} so {{x}} will become {x}
   - Integrate Whoops debug console for development